moved blob storage class to new folder
