+++ src/MarketingPageAcceptanceTests.TestData/Azure/AzureBlobStorage.cs
@@ -1,13 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Text;
 using System.Threading.Tasks;
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
 using FluentAssertions;
 
-namespace MarketingPageAcceptanceTests.TestData.Utils
+namespace MarketingPageAcceptanceTests.TestData.Azure
 {
     public sealed class AzureBlobStorage
     {
